+++ src/stories/states.stories.tsx
@@ -1,6 +1,5 @@
 import React from 'react'
-import { ComponentStory, ComponentMeta } from '@storybook/react'
-import { within, userEvent } from '@storybook/testing-library'
+import { StoryFn, Meta } from '@storybook/react'
 import Component from '../Input'
 import '../../public/styles.css'
 import '../../public/theme/flecto.css'
@@ -11,11 +10,9 @@ export default {
   parameters: {
     layout: 'fullscreen',
   },
-} as ComponentMeta<typeof Component>
+} as Meta<typeof Component>
 
-const Template: ComponentStory<typeof Component> = (args) => (
-  <Component {...args} />
-)
+const Template: StoryFn<typeof Component> = (args) => <Component {...args} />
 
 export const Empty = Template.bind({})